scan-m3u: remake and optimize

+++ packages/xmedia/tools/scan-m3u/package.mk
@@ -22,12 +22,16 @@ pre_configure_target() {
   export LDFLAGS="$LDFLAGS -fPIC"
 }
 
+pre_make_target(){
+  [ -f $ROOT/private/$PKG_NAME/ttv-logo.src ] && cp $ROOT/private/$PKG_NAME/ttv-logo.src $ROOT/$PKG_BUILD
+}
+
 make_target() {
   CC=$CC CFLAGS=$CFLAGS ./shc -v -r -B -f m3u_info.src
   CC=$CC CFLAGS=$CFLAGS ./shc -v -r -B -f scan-m3u.src
   CC=$CC CFLAGS=$CFLAGS ./shc -v -r -B -f live-ttv.src
   CC=$CC CFLAGS=$CFLAGS ./shc -v -r -B -f live-ttv-vdr.src
-  CC=$CC CFLAGS=$CFLAGS ./shc -v -r -B -f ttv-logo.src
+  [ -f ttv-logo.src ] && CC=$CC CFLAGS=$CFLAGS ./shc -v -r -B -f ttv-logo.src
 }
 
 makeinstall_target() {
@@ -37,7 +41,7 @@ makeinstall_target() {
     cp scan-m3u.src.x $INSTALL/usr/bin/scan-m3u
     cp live-ttv.src.x $INSTALL/usr/bin/live-ttv
     cp live-ttv-vdr.src.x $INSTALL/usr/bin/live-ttv-vdr
-    cp ttv-logo.src.x $INSTALL/usr/bin/ttv-logo
+    [ -e ttv-logo.src.x ] && cp ttv-logo.src.x $INSTALL/usr/bin/ttv-logo
   mkdir -p $INSTALL/usr/config/acestream
     cp $PKG_DIR/config/* $INSTALL/usr/config/acestream
 }
